+++ order-service/src/main/java/com/rushcrew/order_service/presentation/dto/request/CreateOrderRequest.java
@@ -37,7 +37,7 @@ public record CreateOrderRequest(
 
 		@NotNull(message = "수량은 필수입니다")
 		@Min(value = 1, message = "수량은 1개 이상이어야 합니다")
-		Integer quantity
+		Long quantity
 	) {}
 
 	public record ShippingInfoRequest(
